improvement: move some fixtures into conftest
